A working height map
Forum goer StarCrusher96, author of Kerbol Star Systems, supplied a Kerbin height map that works. I need to adjust the VertexHeightMap offset to use it as this map is darker than the ScanSat maps. The map is otherwise a faithful Kerbin reproduction, though at a lower resolution. The remaining maps come from Kittopia-dumps.