Use stock textures everywhere
We no longer need external textures or maps to generate these worlds, which should save on runtime memory and download size. This should be ready for production use on KSP 1.2.2, and while it's quite playable on KSP 1.3, we're waiting on biome localization fixes for Kopernicus.

I'll do a (hopefully) final pull request to GregroxMun to hand this project back to him.