Scatterer config, cache changes
Scatterer is a strange beast that writes to its config files on exit. Once I realized that, I figured I just needed to clone Kerbin's config to nodes for Bin, and let Scatterer do the rest.

I also changed the path to the cache files to, what I think is, the intended path of AlienSpacePrograms/Miscellany/Cache/*.bin.