Config cleanup and Kopernicus updates
Going through the configurations since Kopernicus 1.3.0-6's release, which permits us to use BUILTIN maps more than one now. Also moved caches to a common location and eliminated the nested "AlienSpacePrograms" folder.

Also renamed Kerbin to "Bin" when Kerbin is the homeworld. This would permit a user to install a visual add-on, change all references for Kerbin to Bin, and keep their visual add-on settings even if Kerbin is the home world. This was easier than trying to programmatically patch these references depending on whether Kerbin was the homeworld or not. I will have a fork of Stock Visual Enhancements that will support this as well.