Kopernicus 1.5.1-1 tweaks
No major changes. Minor edits to atmospheres to the Atmosphere meters work better. We no longer need the 0.01 m sliver of 1 ATM air to keep rocket ISP values correct as long as there's a staticPressureASL value of 101.325 kPa defined.

The major Kopernicus 1.5.1-1 changes are in the 1.5.1-1-alpha branch created earlier.