Fix #24, Move RA Tylo KSC to Lowlands
I changed the Sun's luminosity values for each home world setting based on Gigantor solar power panel rates found in the stock game. To be extra precise, I used a craft that pointed radial-out and read the highest power value when the panel faced the sun and was parallel to the home world's solar orbit. In Laythe's and Tylo's cases I used the values recorded while orbiting Jool itself.

I also moved the KSC on Homeworld-Tylo to lower ground when RA was enabled. This enables thin-air craft to function. A modified Aeris 4A can get to Tylo orbit and back safely.