Address #21, put back KSC peninsula on Bin
I can't seem to get City{} sections to behave the way I expect them to. I wonder if Kopernicus is building them on the opposite side of the world. But even so, I can put the flat surfaces where the KSC was on Kerbin when Kerbin isn't the home world. So no KSC and KSC2, but at least people can now complete the Bin Space Program challenge.